+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/store/RoomStore.kt
@@ -4,6 +4,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 import net.folivo.trixnity.client.MatrixClientConfiguration
 import net.folivo.trixnity.client.store.cache.FullRepositoryObservableCache
+import net.folivo.trixnity.client.store.cache.ObservableCacheStatisticCollector
 import net.folivo.trixnity.client.store.repository.RepositoryTransactionManager
 import net.folivo.trixnity.client.store.repository.RoomRepository
 import net.folivo.trixnity.core.model.RoomId
@@ -11,11 +12,13 @@ import net.folivo.trixnity.core.model.RoomId
 class RoomStore(
     roomRepository: RoomRepository,
     tm: RepositoryTransactionManager,
-    storeScope: CoroutineScope,
     config: MatrixClientConfiguration,
+    statisticCollector: ObservableCacheStatisticCollector,
+    storeScope: CoroutineScope,
 ) : Store {
     private val roomCache =
         FullRepositoryObservableCache(roomRepository, tm, storeScope, config.cacheExpireDurations.room) { it.roomId }
+            .also(statisticCollector::addCache)
 
     override suspend fun clearCache() = deleteAll()