+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/store/cache/ObservableCache.kt
@@ -2,6 +2,7 @@ package net.folivo.trixnity.client.store.cache
 
 import io.github.oshai.kotlinlogging.KotlinLogging
 import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.cancel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
@@ -46,7 +47,7 @@ internal interface ObservableCacheStore<K, V> {
 /**
  * An index to track which entries have been added to or removed from the cache.
  */
-internal interface ObservableMapIndex<K> {
+internal interface ObservableCacheIndex<K> {
     /**
      * Called, when an entry is added to the cache.
      */
@@ -68,6 +69,8 @@ internal interface ObservableMapIndex<K> {
      * Get the subscription count on an index entry, which uses an entry of the cache.
      */
     suspend fun getSubscriptionCount(key: K): Flow<Int>
+
+    suspend fun collectStatistic(): ObservableCacheIndexStatistic?
 }
 
 /**
@@ -149,7 +152,7 @@ internal open class ObservableCacheBase<K, V>(
         addIndex(RemoverJobExecutingIndex(name, values, cacheScope, expireDuration))
     }
 
-    fun addIndex(index: ObservableMapIndex<K>) {
+    fun addIndex(index: ObservableCacheIndex<K>) {
         values.indexes.update { it + index }
     }
 
@@ -197,6 +200,19 @@ internal open class ObservableCacheBase<K, V>(
             }
         }
     }
+
+    internal suspend fun collectStatistic(): ObservableCacheStatistic {
+        val (all, subscribed) = values.internalRead {
+            count() to values.count { it.subscriptionCount.value > 0 }
+        }
+        return ObservableCacheStatistic(
+            name = name,
+            all = all,
+            subscribed = subscribed,
+            indexes = values.indexes.value.mapNotNull { it.collectStatistic() }
+        )
+    }
+
 }
 
 private class RemoverJobExecutingIndex<K, V>(
@@ -204,7 +220,7 @@ private class RemoverJobExecutingIndex<K, V>(
     private val values: ConcurrentObservableMap<K, MutableStateFlow<CacheValue<V?>>>,
     private val cacheScope: CoroutineScope,
     private val expireDuration: Duration = 1.minutes,
-) : ObservableMapIndex<K> {
+) : ObservableCacheIndex<K> {
     private val infiniteCache = expireDuration.isInfinite()
     override suspend fun onPut(key: K) {
         if (infiniteCache.not()) {
@@ -225,6 +241,7 @@ private class RemoverJobExecutingIndex<K, V>(
                     if (subscriptionCount == 0 && (stale || indexSubscriptionCount == 0)) {
                         log.trace { "$name: remove value from cache with key $key" }
                         values.remove(key, stale)
+                        cancel("key $key removed from cache")
                     }
                 }
             }
@@ -233,5 +250,8 @@ private class RemoverJobExecutingIndex<K, V>(
 
     override suspend fun onRemove(key: K, stale: Boolean) {}
     override suspend fun onRemoveAll() {}
-    override suspend fun getSubscriptionCount(key: K): StateFlow<Int> = MutableStateFlow(0)
+    override suspend fun collectStatistic(): ObservableCacheIndexStatistic? = null
+
+    private val zeroStateFlow = MutableStateFlow(0)
+    override suspend fun getSubscriptionCount(key: K): StateFlow<Int> = zeroStateFlow
 }