+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/store/createStoreModule.kt
@@ -1,12 +1,19 @@
 package net.folivo.trixnity.client.store
 
 import net.folivo.trixnity.client.media.MediaStore
+import net.folivo.trixnity.client.store.cache.ObservableCacheStatisticCollector
+import net.folivo.trixnity.core.EventHandler
+import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.bind
 import org.koin.dsl.module
 
 fun createStoreModule() = module {
     singleOf(::TransactionManagerImpl).bind<TransactionManager>()
+    singleOf(::ObservableCacheStatisticCollector) {
+        bind<EventHandler>()
+        bind<ObservableCacheStatisticCollector>()
+    }
     singleOf(::AccountStore)
     singleOf(::ServerDataStore)
     singleOf(::GlobalAccountDataStore)
@@ -22,6 +29,7 @@ fun createStoreModule() = module {
             roomKeyRequestRepository = get(),
             tm = get(),
             config = get(),
+            statisticCollector = get(),
             storeScope = get()
         )
     }