+++ trixnity-client/src/commonTest/kotlin/net/folivo/trixnity/client/testutils.kt
@@ -9,6 +9,7 @@ import kotlinx.coroutines.launch
 import kotlinx.serialization.json.Json
 import net.folivo.trixnity.client.mocks.RepositoryTransactionManagerMock
 import net.folivo.trixnity.client.store.*
+import net.folivo.trixnity.client.store.cache.ObservableCacheStatisticCollector
 import net.folivo.trixnity.client.store.repository.*
 import net.folivo.trixnity.clientserverapi.client.MatrixClientServerApiClientImpl
 import net.folivo.trixnity.clientserverapi.model.media.GetMediaConfig
@@ -47,6 +48,7 @@ fun mockMatrixClientServerApiClient(
 suspend fun getInMemoryAccountStore(scope: CoroutineScope) = AccountStore(
     InMemoryAccountRepository(),
     RepositoryTransactionManagerMock(),
+    ObservableCacheStatisticCollector(),
     scope
 ).apply {
     init()
@@ -61,6 +63,7 @@ suspend fun getInMemoryServerDataStore(scope: CoroutineScope) = ServerDataStore(
         )
     },
     RepositoryTransactionManagerMock(),
+    ObservableCacheStatisticCollector(),
     scope
 ).apply { init() }
 
@@ -69,6 +72,7 @@ suspend fun getInMemoryRoomAccountDataStore(scope: CoroutineScope) = RoomAccount
     RepositoryTransactionManagerMock(),
     DefaultEventContentSerializerMappings,
     MatrixClientConfiguration(),
+    ObservableCacheStatisticCollector(),
     scope
 ).apply { init() }
 
@@ -77,6 +81,7 @@ suspend fun getInMemoryGlobalAccountDataStore(scope: CoroutineScope) = GlobalAcc
     RepositoryTransactionManagerMock(),
     DefaultEventContentSerializerMappings,
     MatrixClientConfiguration(),
+    ObservableCacheStatisticCollector(),
     scope
 ).apply { init() }
 
@@ -89,6 +94,7 @@ suspend fun getInMemoryOlmStore(scope: CoroutineScope) = OlmCryptoStore(
     InMemoryOutboundMegolmSessionRepository(),
     RepositoryTransactionManagerMock(),
     MatrixClientConfiguration(),
+    ObservableCacheStatisticCollector(),
     scope
 ).apply { init() }
 
@@ -103,14 +109,16 @@ suspend fun getInMemoryKeyStore(scope: CoroutineScope) = KeyStore(
     InMemoryRoomKeyRequestRepository(),
     RepositoryTransactionManagerMock(),
     MatrixClientConfiguration(),
+    ObservableCacheStatisticCollector(),
     scope
 ).apply { init() }
 
 suspend fun getInMemoryRoomStore(scope: CoroutineScope) = RoomStore(
     InMemoryRoomRepository(),
     RepositoryTransactionManagerMock(),
-    scope,
     MatrixClientConfiguration(),
+    ObservableCacheStatisticCollector(),
+    scope,
 ).apply { init() }
 
 suspend fun getInMemoryRoomTimelineStore(scope: CoroutineScope) = RoomTimelineStore(
@@ -118,6 +126,7 @@ suspend fun getInMemoryRoomTimelineStore(scope: CoroutineScope) = RoomTimelineSt
     InMemoryTimelineEventRelationRepository(),
     RepositoryTransactionManagerMock(),
     MatrixClientConfiguration(),
+    ObservableCacheStatisticCollector(),
     scope
 ).apply { init() }
 
@@ -126,6 +135,7 @@ suspend fun getInMemoryRoomStateStore(scope: CoroutineScope) = RoomStateStore(
     RepositoryTransactionManagerMock(),
     DefaultEventContentSerializerMappings,
     MatrixClientConfiguration(),
+    ObservableCacheStatisticCollector(),
     scope
 ).apply { init() }
 
@@ -134,6 +144,7 @@ suspend fun getInMemoryRoomUserStore(scope: CoroutineScope) = RoomUserStore(
     InMemoryRoomUserReceiptsRepository(),
     RepositoryTransactionManagerMock(),
     MatrixClientConfiguration(),
+    ObservableCacheStatisticCollector(),
     scope
 ).apply { init() }
 
@@ -141,14 +152,16 @@ suspend fun getInMemoryMediaCacheMapping(scope: CoroutineScope) = MediaCacheMapp
     InMemoryMediaCacheMappingRepository(),
     RepositoryTransactionManagerMock(),
     MatrixClientConfiguration(),
+    ObservableCacheStatisticCollector(),
     scope
 ).apply { init() }
 
 suspend fun getInMemoryRoomOutboxMessageStore(scope: CoroutineScope) = RoomOutboxMessageStore(
     InMemoryRoomOutboxMessageRepository(),
     RepositoryTransactionManagerMock(),
-    scope,
     MatrixClientConfiguration(),
+    ObservableCacheStatisticCollector(),
+    scope,
 ).apply { init() }
 
 suspend fun ShouldSpecContainerScope.clearOutdatedKeys(keyStoreBuilder: () -> KeyStore) {